added finance calculator

+++ python_introduction/finance_calculator.py
@@ -3,7 +3,7 @@ total_expense = int(input("Enter your monthly expense: "))
 
 monthly_saving = monthly_income - total_expense
 
-projected_savings = ( monthly_saving * 12 +(monthly_saving *12 *0.05))
+projected_savings = (monthly_saving * 12) +  float( monthly_saving *12 *0.05)
 
 print(monthly_saving)
 print(projected_savings)
